requirements.txt
tornado==4.0.2
torndb==0.3
MySQL-python==1.2.5
PyYAML==3.11
pycrypto>=2.6.1
redis==2.10.3
17MonIP==0.2.0
ua-parser==0.3.5
user-agents==0.3.0
requests==2.2.1
ujson==1.33
